+++ docs/changelog.md
@@ -62,6 +62,10 @@
 
 ## 更新说明详情
 
+1.10.43 rem: rem index_vix interface
+
+    1. 移除 index_vix 接口
+
 1.10.42 fix: fix fund_rating_all interface
 
     1. 修复 fund_rating_all 接口
@@ -2575,6 +2579,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.43 rem: rem index_vix interface
 
 1.10.42 fix: fix fund_rating_all interface